Create docker-compose.yml
It should be noted that 8081 & 8441 are ports which have been arbitrary chosen & can be anything valid. When when finds the IP address using docker network inspect drupal_default the webpage can be accessed using 172.20.0.1:8081 as the IP address of drupal_default in my case was 172.20.0.1. In case the program is being run on Ubuntu 18.04 just replace 8080 with 80 & 8443 with 443 above. I have already tested it and find it strange why one need to make these changes. It appears the tcp port defaults have been changed due to some reason.